+++ libs/ofxESCPOS/Commands/src/DefaultMechanismCommands.cpp
@@ -51,7 +51,11 @@ std::size_t DefaultMechanismCommands::cut(BaseCodes::PaperCut type,
     }
     else
     {
-        const uint8_t command[4] = { BaseCodes::GS, 'V', type + 'A', numDotsToFeed};
+        const uint8_t command[4] = { BaseCodes::GS,
+									'V',
+									 static_cast<uint8_t>(type + 'A'),
+									numDotsToFeed};
+
         return writeBytes(command, 4); // shift -> A/B
     }
 }
